Starting to move miriad tests to separate file
